Exports the solution matrix to the final solution.

Exports the solution matrix to the final solution.

processSolution(
  solutionMatrix,
  variables,
  dataPreprocessed,
  carnivalOptions,
  newDataRepresentation = TRUE
)

Arguments

solutionMatrix

the output matrix from ILP solver containing variables list (rows) and their values in different solutions (columns).

variables

list of nodes, edges and measurements variables generated by createLpFormulation_v2.

dataPreprocessed

list containing preprocessed priorKnowledgeNetwork, measurements, weights (if provided), perturbations (if provided).

carnivalOptions

all options of CARNIVAL.

newDataRepresentation

TRUE by default. For debugging with the old data representation, put to FALSE.

Value

Carnival results exported from the solution matrix. see runCARNIVAL for details.